+++ crates/crab-registry/src/lib.rs
@@ -20,15 +20,18 @@ mod blocks_1_20_3;
 mod blocks_1_20_5;
 mod blocks_1_21;
 mod blocks_1_21_3;
+mod blocks_1_21_4;
 mod entities_1_20_1;
 mod entities_1_20_3;
 mod entities_1_20_5;
 mod entities_1_21_3;
+mod entities_1_21_4;
 mod items_1_20_1;
 mod items_1_20_3;
 mod items_1_20_5;
 mod items_1_21;
 mod items_1_21_3;
+mod items_1_21_4;
 
 pub use blocks_1_20_1::BLOCKS_1_20_1;
 pub use blocks_1_20_2::BLOCKS_1_20_2;
@@ -36,15 +39,18 @@ pub use blocks_1_20_3::BLOCKS_1_20_3;
 pub use blocks_1_20_5::BLOCKS_1_20_5;
 pub use blocks_1_21::BLOCKS_1_21;
 pub use blocks_1_21_3::BLOCKS_1_21_3;
+pub use blocks_1_21_4::BLOCKS_1_21_4;
 pub use entities_1_20_1::ENTITIES_1_20_1;
 pub use entities_1_20_3::ENTITIES_1_20_3;
 pub use entities_1_20_5::ENTITIES_1_20_5;
 pub use entities_1_21_3::ENTITIES_1_21_3;
+pub use entities_1_21_4::ENTITIES_1_21_4;
 pub use items_1_20_1::ITEMS_1_20_1;
 pub use items_1_20_3::ITEMS_1_20_3;
 pub use items_1_20_5::ITEMS_1_20_5;
 pub use items_1_21::ITEMS_1_21;
 pub use items_1_21_3::ITEMS_1_21_3;
+pub use items_1_21_4::ITEMS_1_21_4;
 
 static REGISTRY_PROFILE: AtomicU8 = AtomicU8::new(0);
 
@@ -58,6 +64,7 @@ pub fn set_protocol(protocol: i32) {
             766 => 3,
             767 => 4,
             768 => 5,
+            769 => 6,
             _ => 0,
         },
         Ordering::Relaxed,
@@ -73,6 +80,7 @@ pub fn blocks() -> &'static [BlockDef] {
         3 => BLOCKS_1_20_5,
         4 => BLOCKS_1_21,
         5 => BLOCKS_1_21_3,
+        6 => BLOCKS_1_21_4,
         _ => BLOCKS_1_20_1,
     }
 }
@@ -85,6 +93,7 @@ pub fn items() -> &'static [ItemDef] {
         3 => ITEMS_1_20_5,
         4 => ITEMS_1_21,
         5 => ITEMS_1_21_3,
+        6 => ITEMS_1_21_4,
         _ => ITEMS_1_20_1,
     }
 }
@@ -97,6 +106,7 @@ pub fn entities() -> &'static [EntityDef] {
         3 => ENTITIES_1_20_5,
         4 => ENTITIES_1_20_5,
         5 => ENTITIES_1_21_3,
+        6 => ENTITIES_1_21_4,
         _ => ENTITIES_1_20_1,
     }
 }
@@ -1116,5 +1126,29 @@ mod tests {
             .find(|entity| entity.name == "creaking")
             .unwrap();
         assert_eq!(creaking.id, 29);
+
+        let mace_769 = ITEMS_1_21_4
+            .iter()
+            .find(|item| item.name == "mace")
+            .unwrap();
+        assert_eq!(mace_769.id, 1144);
+        let resin = ITEMS_1_21_4
+            .iter()
+            .find(|item| item.name == "resin_clump")
+            .unwrap();
+        assert_eq!(resin.id, 376);
+        let creaking_heart = BLOCKS_1_21_4
+            .iter()
+            .find(|block| block.name == "minecraft:creaking_heart")
+            .unwrap();
+        assert_eq!(creaking_heart.default_state, 2926);
+        assert_eq!(
+            ENTITIES_1_21_4
+                .iter()
+                .find(|entity| entity.name == "creaking")
+                .unwrap()
+                .id,
+            29
+        );
     }
 }